dom: use domtree for bb_dominates()
The function bb_dominates() do a dominance query
on two BB but do this by walking the CFG tree.
Now that we have the dominance tree, we don't need anymore
to do this walk, we can use the dom tree itself.

This patch just replace the calls bb_dominates() by
domtree_dominates() since they have a slightly different
interface.

Note: in the next version, it'll be better/simpler to rename
      domtree_dominates() by bb_dominates().
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>