doc: the man page in reST
Manpages are a good thing but as an input format it's a
nightmare. It would be nice to have a more adapted format
for it.

So convert the manpage to reST and add to the Makefile what
is needed to generate the man page from it.

NB. This is still an experiment, but it's my intention to
    make this the source file for the manpage.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>