doc: fix weirdness with option lists
sphinx considers an option '-fsomething' as the option 'f' with
parameter 'something'. The option part '-f' is then displayed
in bold and the argument part 'something' is displayed in italic.

This is often the right interpretation and look quite good for
some styles/options, like '-lsparse' but I find it quite strange
looking for GCC/sparse options like '-Wunknown-attribute' and
even more for '-fmax-error=1000' where the real argument is '1000',
not 'max-error=1000', especially on the man-page.

Make the man-page better looking with a little bit of sed magic.

Note: this is not a serious fix, it's more as experiment to be
      able to see the expected output of the man-page.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>