fix dominance testing of mixed types
The function dominates() is used to test the dominance of
one instruction over another one. To do this different things
need to be checked, like:
- same symbol or not
- same address & same offset
- partial overlapping

One case is not covered though: when type cohercion is done
via a pointer to an union (IOW when a memory location is used
to store one type and is then read back as another type).
For example, the location is first stored as a float and then is
read back as an integer of the same size. Currently, sparse
will consider that the store effectively dominates the load
which will then allow to simplify away the load and use the
pseudo holding the float value for the expected integer.

There is surely several ways to fix this problem.
The solution used in this patch is to check if both instructions
are of of the same floating-pointness.

Note: this solution is probably incomplete (what about compound
      types, for example?)

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>