kds: shortcut for kill_dead_stores()
In kill_dead_stores(), instructions must be scanned backward
in order to check if the store is dead and can be removed.

However, if the corresponding pseudo has a single user,
this user is a store and the it correspond to a local symbol,
then it's not needed to scan, the store can be directly be removed.

This is somehow important for some semi-pathological case where
some BB contains a lot of dead stores because the current way is
annoyingly O(n^2).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>